+++ src/line_style.rs
@@ -10,7 +10,7 @@ use {
 /// It's made of
 ///  - the base style of the compounds
 ///  - the alignment
-#[derive(Default, Clone, Debug)]
+#[derive(Default, Clone, Debug, PartialEq)]
 pub struct LineStyle {
     pub compound_style: CompoundStyle,
     pub align: Alignment,