+++ src/parse/parse_compound_style.rs
@@ -1,9 +1,22 @@
 use {
     super::*,
-    crate::CompoundStyle,
+    crate::{
+        ATTRIBUTES,
+        CompoundStyle,
+    },
 };
 
 /// Read a Minimad CompoundStyle from a string.
+///
+/// May contain attributes and from 0 to 2 colors, the first encountered
+/// one being the foreground.
+///
+/// Examples:
+/// * `#ab00c1 strikeout`
+/// * `"red yellow bold italic"`
+/// * `""`
+/// * `"gray(2) gray(20)"`
+/// * `""`
 pub fn parse_compound_style(s: &str) -> Result<CompoundStyle, ParseStyleTokenError> {
     let tokens = parse_style_tokens(s)?;
     Ok(tokens.as_slice().into())
@@ -42,5 +55,23 @@ impl From<&[StyleToken]> for CompoundStyle {
             }
         }
         style
+    }
+}
+
+impl PushStyleTokens for CompoundStyle {
+    fn push_style_tokens(&self, tokens: &mut Vec<StyleToken>) {
+        if let Some(fg) = self.get_fg() {
+            tokens.push(StyleToken::Color(fg));
+        } else if self.get_bg().is_some() {
+            tokens.push(StyleToken::None);
+        }
+        if let Some(bg) = self.get_bg() {
+            tokens.push(StyleToken::Color(bg));
+        }
+        for &attr in ATTRIBUTES {
+            if self.has_attr(attr) {
+                tokens.push(StyleToken::Attribute(attr));
+            }
+        }
     }
 }