+++ src/serde/serde_line_style.rs
@@ -2,8 +2,9 @@ use {
     crate::{
         LineStyle,
         parse_line_style,
+        parse::PushStyleTokens,
     },
-    serde::de,
+    serde::{de, Serialize, Serializer},
 };
 
 impl<'de> de::Deserialize<'de> for LineStyle {
@@ -13,5 +14,14 @@ impl<'de> de::Deserialize<'de> for LineStyle {
     {
         let s = String::deserialize(deserializer)?;
         parse_line_style(&s).map_err(de::Error::custom)
+    }
+}
+
+impl Serialize for LineStyle {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        serializer.serialize_str(&self.to_style_tokens_string())
     }
 }